Export CMake targets
These are the smallest required changes to export targets for PIO_C,
PIO_Fortran and PIO_GPTL targets.

Note that with this change, the installed PIO package is not
relocatable. That is, target properties which may contain paths, such as
INTERFACE_INCLUDE_DIRECTORIES and INTERFACE_LINK_LIBRARIES of the
exported targets may contain the absolute paths to libraries and include
directories of external dependencies as found on the machine the package
was made on. If any of these absolute paths are no longer valid on the
machine the package is installed on, then the installation will break.
See [Creating Relocatable Packages][creating-relocatable-packages] for
more details.

[creating-relocatable-packages]: https://cmake.org/cmake/help/latest/manual/cmake-packages.7.html#creating-relocatable-packages